Modified User Handling

+++ webserver/src/main/java/org/example/webserver/components/subtask/SubtaskRepository.java
@@ -2,5 +2,5 @@ package org.example.webserver.components.subtask;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubtaskRepository extends JpaRepository<SubtaskModel, Long> {
+public interface SubtaskRepository extends JpaRepository<SubtaskModel, Integer> {
 }
